+++ src/models/bgp/attributes/mod.rs
@@ -476,6 +476,10 @@ pub enum AttributeValue {
     Clusters(Vec<u32>),
     MpReachNlri(Nlri),
     MpUnreachNlri(Nlri),
+    /// BGP Link-State attribute - RFC 7752
+    LinkState(crate::models::bgp::linkstate::LinkStateAttribute),
+    /// BGP Tunnel Encapsulation attribute - RFC 9012
+    TunnelEncapsulation(crate::models::bgp::tunnel_encap::TunnelEncapAttribute),
     Development(Vec<u8>),
     Deprecated(AttrRaw),
     Unknown(AttrRaw),
@@ -532,6 +536,8 @@ impl AttributeValue {
             AttributeValue::Clusters(_) => AttrType::CLUSTER_LIST,
             AttributeValue::MpReachNlri(_) => AttrType::MP_REACHABLE_NLRI,
             AttributeValue::MpUnreachNlri(_) => AttrType::MP_UNREACHABLE_NLRI,
+            AttributeValue::LinkState(_) => AttrType::BGP_LS_ATTRIBUTE,
+            AttributeValue::TunnelEncapsulation(_) => AttrType::TUNNEL_ENCAPSULATION,
             AttributeValue::Development(_) => AttrType::DEVELOPMENT,
             AttributeValue::Deprecated(x) | AttributeValue::Unknown(x) => x.attr_type,
         }
@@ -558,6 +564,7 @@ impl AttributeValue {
             AttributeValue::Clusters(_) => Some(OptionalNonTransitive),
             AttributeValue::MpReachNlri(_) => Some(OptionalNonTransitive),
             AttributeValue::MpUnreachNlri(_) => Some(OptionalNonTransitive),
+            AttributeValue::LinkState(_) => Some(OptionalNonTransitive),
             _ => None,
         }
     }