+++ src/models/bgp/mod.rs
@@ -5,13 +5,17 @@ pub mod capabilities;
 pub mod community;
 pub mod elem;
 pub mod error;
+pub mod linkstate;
 pub mod role;
+pub mod tunnel_encap;
 
 pub use attributes::*;
 pub use community::*;
 pub use elem::*;
 pub use error::*;
+pub use linkstate::*;
 pub use role::*;
+pub use tunnel_encap::*;
 
 use crate::models::network::*;
 use capabilities::{