+++ src/parser/bgp/messages.rs
@@ -485,6 +485,7 @@ mod tests {
             safi: Safi::Unicast,
             next_hop: None,
             prefixes: vec![],
+            link_state_nlris: None,
         })]);
         let msg = BgpUpdateMessage {
             withdrawn_prefixes: vec![],
@@ -519,6 +520,7 @@ mod tests {
             safi: Safi::Unicast,
             next_hop: None,
             prefixes: vec![],
+            link_state_nlris: None,
         })]);
         let msg = BgpUpdateMessage {
             withdrawn_prefixes: vec![],
@@ -533,6 +535,7 @@ mod tests {
             safi: Safi::Unicast,
             next_hop: None,
             prefixes: vec![prefix],
+            link_state_nlris: None,
         })]);
         let msg = BgpUpdateMessage {
             withdrawn_prefixes: vec![],
@@ -547,6 +550,7 @@ mod tests {
             safi: Safi::Unicast,
             next_hop: None,
             prefixes: vec![prefix],
+            link_state_nlris: None,
         })]);
         let msg = BgpUpdateMessage {
             withdrawn_prefixes: vec![],
@@ -562,6 +566,7 @@ mod tests {
                 safi: Safi::Unicast,
                 next_hop: None,
                 prefixes: vec![],
+                link_state_nlris: None,
             }),
             AttributeValue::AtomicAggregate,
         ]);